requirements.txt
# Core dependencies for video processing and scene detection
opencv-python==4.12.0.88
numpy==2.2.6

# Google Cloud and AI services
google-cloud-videointelligence==2.16.2
google-genai

# FAL AI services
fal-client==0.7.0

# Environment and configuration
python-dotenv==1.1.1

# FastAPI backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
debugpy==1.8.0

# Additional dependencies that might be needed
requests==2.32.5
pathlib2==2.3.7

python-multipart
